pyproject.toml
[project]
name = "fastapi-chat-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Dafoe17",email = "ram.and.rem207@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B"]
ignore = ["B008"]
exclude = ["alembic"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = false
